will not be uninstalled after test run

+++ test/functional/setup.js
@@ -13,9 +13,15 @@ global.shell = (command, options) =>
 // todo create custom chai assertion
 // todo think of something for process.cwd to return file dir during tests
 
+// todo restore initial version and not just leave fresh install.
+// At least if it's failed
+let wasInstalledBeforeRun
+
 before(function () {
   this.timeout(10000)
 
+  wasInstalledBeforeRun = shell(`npm list -g -depth ${name}`)
+    .stdout.includes(name)
   installGlobalFromPack()
   process.chdir(__dirname)
 })
@@ -24,5 +30,5 @@ after(function () {
   this.timeout(10000)
 
   process.chdir(rootFolder)
-  shell(`npm uninstall -g ${name}`)
+  if (!wasInstalledBeforeRun) shell(`npm uninstall -g ${name}`)
 })
